test(ai): make the action_intents immutability contract run in CI
The rewritten contract test covered the new approval_scope /
classification_version / effect_digest columns, but its ~28 live-DB cases
never executed anywhere in CI: they sat behind describe.runIf(DATABASE_URL),
the blocking test-api job has no Postgres, and the file was not in the
integration config's include list. The immutability of approval_scope — the
property that stops an intent switching approval classification after
approvers have acted — was asserted nowhere.

Split the suite along the line CI actually draws:
- Static DDL/parse assertions stay in src/db/migration-action-intents.test.ts
  and keep running in the fast, blocking test-api job.
- The 28 behavioral cases move to
  src/__tests__/integration/actionIntentsImmutabilityTrigger.integration.test.ts
  with per-test fixtures so they are truncation-safe. That directory is matched
  wholesale by both configs (integration include, unit exclude), so it needs no
  hand-maintained entry in the list that has gone stale before.
- Shared deny-list derivation extracted to testUtils so the two suites cannot
  drift into vacuity independently.

Verified non-vacuous: removing the approval_scope clause from the trigger in a
throwaway test DB fails the suite twice over — the behavioral UPDATE case and
the live-function-vs-migration drift gate — and restoring returns it to 28/28.

Runs in the Integration Tests job, which is blocking on PRs and main (the
continue-on-error in ci.yml belongs to smoke-test, not integration-test).
